Detect and fix erroneous HW pointer in updateIRQJPEG_wp
It looks like compessor is reporting erroneous HW pointer when last
frame ends precisely at the end of buffer and 32 zero bytes start from
the beginning of the buffer. HW pointer in this case should be 0x20, but
it is 0x00 in fact. This situation is detected and fixed in updateIRQJPEG_wp.